+++ .config/dotnet-tools.json
@@ -1,13 +1,36 @@
 {
-  "version": 1,
-  "isRoot": true,
-  "tools": {
-    "dotnet-mgcb": {
-      "version": "3.8.4.1",
-      "commands": [
-        "mgcb"
-      ],
-      "rollForward": false
+    "version": 1,
+    "isRoot": true,
+    "tools": {
+        "dotnet-mgcb": {
+            "version": "3.8.4.1",
+            "commands": [
+                "mgcb"
+            ]
+        },
+        "dotnet-mgcb-editor": {
+            "version": "3.8.4.1",
+            "commands": [
+                "mgcb-editor"
+            ]
+        },
+        "dotnet-mgcb-editor-linux": {
+            "version": "3.8.4.1",
+            "commands": [
+                "mgcb-editor-linux"
+            ]
+        },
+        "dotnet-mgcb-editor-windows": {
+            "version": "3.8.4.1",
+            "commands": [
+                "mgcb-editor-windows"
+            ]
+        },
+        "dotnet-mgcb-editor-mac": {
+            "version": "3.8.4.1",
+            "commands": [
+                "mgcb-editor-mac"
+            ]
+        }
     }
-  }
 }
