+++ Source/MouseInfo.cs
@@ -2,7 +2,7 @@
 using Microsoft.Xna.Framework.Input;
 
 
-namespace IntroToMonoGame;
+namespace MagicBrosMario.Source;
 
 public class MouseInfo : IController
 {
