+++ Source/MovingSprite.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
-namespace IntroToMonoGame;
+namespace MagicBrosMario.Source;
 
 public class MovingSprite : ISprite
 {
